Add more functionality to the TextBox class.
Added support for navigation with left/right arrow keys.
Backspace and delete keys work as expected.
Clicking on text will move the caret as expected.
Removed the Anchor enum (it was a hack) and added a new DrawText
overload that allows text to be clipped.
Added some functions to the Font class for retrieving glyph info.