Get the Dialog class working.
Dialog now takes a dimentions param instead of a rect.
Added a Panel control that is used to hold any controls on the Dialog.
Added supporting APIs for adding controls to the Dialog.
Added OnLocationChanged to Control, providing an extensibility point
for derived classes.  Whenever a Control's location is changed the
delta of X and Y will be provided.  This is being used by the Panel and
Dialog classes enabling their constituent controls to move with them.
Moved the Dimentions struct into its own header file allowing it to
be used by classes other than Window.
Replaced some "magic numbers" in the Dialog class with constants.
Adding a control to a panel that's hidden will hide the control.